correct mistakes
