+++ intensive/flutter-task-4/lib/constants/app_strings.dart
@@ -1,0 +1,15 @@
+class AppStrings {
+  static const appTitle = 'Это тайтл';
+  static const checkNumber = 'Чек № 73';
+  static const checkDate = '15.02.24 в 00:24';
+  static const productListTitle = 'Список покупок';
+  static const nothingHereYet = 'Здесь пока ничего нет';
+  static const inYourPurchase = 'В вашей покупке';
+  static const totalItems = 'товаров';
+  static const discount = 'Скидка';
+  static const total = 'Итого';
+  static const catalog = 'Каталог';
+  static const search = 'Поиск';
+  static const basket = 'Корзина';
+  static const personal = 'Личное';
+}
