+++ intensive/flutter-task-4/lib/models/sale_info.dart
@@ -1,0 +1,13 @@
+class SaleInfo {
+  final int total;
+  final int totalSalePercent;
+  final int totalSaleRub;
+  final int totalPriceWithSale;
+
+  SaleInfo({
+    required this.total,
+    required this.totalSalePercent,
+    required this.totalSaleRub,
+    required this.totalPriceWithSale,
+  });
+}
